requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask_migrate==3.1.0
flask_cors==3.0.10
gunicorn==20.1.0
requests==2.26.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1